backend/setapp/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
psycopg2-binary
Pillow
python-dotenv
